+++ scripts/asr_language_modelling/ngram_lm/train_kenlm.py
@@ -14,38 +14,141 @@
 #
 
 import argparse
+import gc
+import json
 import logging
 import os
 import subprocess
 import sys
 
-parser = argparse.ArgumentParser(
-    description='Train an n-gram language model with KenLM to be used with BeamSearch decoder of ASR models.'
-)
-parser.add_argument("--input_path", required=True, type=str)
-parser.add_argument("--output_path", required=True, type=str)
-parser.add_argument("--ngram_length", required=True, type=int)
-parser.add_argument("--kenlm_path", required=True, type=str)
+from joblib import Parallel, delayed
+from tqdm.auto import tqdm
 
-args = parser.parse_args()
 
-""" LMPLZ ARGUMENT SETUP """
-args = [
-    os.path.join(args.kenlm_path, 'lmplz'),
-    "-o",
-    args.ngram_length,
-    "--text",
-    args.input_path,
-    "--arpa",
-    f"{args.output_path}.arpa",
-    "--discount_fallback",
-]
+def read_text(path, lowercase: bool = False):
+    lines_read = 0
+    text_dataset = []
 
-result = subprocess.run(args, capture_output=False, text=True, stdout=sys.stdout, stderr=sys.stderr)
+    # Roughly 40 Million lines of text
+    with open(path, 'r') as f:
+        reader = tqdm(iter(lambda: f.readline(), ''), desc="Read 0 lines", unit=' lines')
+        for i, line in enumerate(reader):
+            # Clean text line
+            # print(path)
+            if path.endswith('.json'):
+                line = json.loads(line)['text']
+                # print(line)
 
-""" BINARY BUILD """
-args = [os.path.join(args.kenlm_path, "build_binary"), "trie", f"{args.output_path}.arpa", args.output_path]
+            line = line.replace("\n", "").strip()
+            if lowercase:
+                line = line.lower()
 
-logging.info(f"Running binary_build command \n\n{' '.join(args)}\n\n")
+            if line:
+                text_dataset.append(line)
 
-result = subprocess.run(args, capture_output=False, text=True, stdout=sys.stdout, stderr=sys.stderr)
+                lines_read += 1
+                if lines_read % 100000 == 0:
+                    reader.set_description(f"Read {lines_read} lines")
+
+    return text_dataset
+
+
+def tokenize_str(texts, tokenizer, ids_to_tokens, offset):
+    tokenized_text = []
+    for text in texts:
+        tok_text = tokenizer.text_to_ids(text)
+        tok_text = [chr(token + offset) for token in tok_text]
+        # tok_text = [ids_to_tokens[idx] for idx in tok_text]
+        tokenized_text.append(tok_text)
+    return tokenized_text
+
+
+def tokenize_text(data, tokenizer, path, chunk_size=8192, buffer_size=32, token_offset=100):
+    dataset_len = len(data)
+    print(
+        f"Chunking {dataset_len} rows into {dataset_len / float(chunk_size):0.4f} tasks (each chunk contains {chunk_size} elements)"
+    )
+
+    vocabulary = tokenizer.tokenizer.get_vocab()
+    idx_to_token = {v: k for k, v in vocabulary.items()}
+
+    current_step = 0
+
+    if os.path.exists(path):
+        print(f"Deleting previous file : {path}")
+        os.remove(path)
+
+    with Parallel(n_jobs=-2, verbose=10) as parallel:
+        while True:
+            start = current_step * chunk_size
+            end = min((current_step + buffer_size) * chunk_size, dataset_len)
+
+            tokenized_data = parallel(
+                delayed(tokenize_str)(data[start : start + chunk_size], tokenizer, idx_to_token, token_offset)
+                for start in range(start, end, chunk_size)
+            )
+
+            # Write dataset
+            write_dataset(tokenized_data, path)
+
+            current_step += len(tokenized_data)
+
+            print(f"Finished writing {len(tokenized_data)} chunks to {path}. Current chunk index = {current_step}")
+
+            del tokenized_data
+            gc.collect()
+
+            if end >= dataset_len:
+                break
+
+
+def write_dataset(chunks, path):
+    basedir = os.path.dirname(path)
+
+    if not os.path.exists(basedir):
+        os.makedirs(basedir, exist_ok=True)
+
+    with open(path, 'a+', encoding='utf-8') as f:
+        for chunk_idx in tqdm(range(len(chunks)), desc='Chunk ', total=len(chunks), unit=' chunks'):
+            for text in chunks[chunk_idx]:
+                line = ' '.join(text)
+                f.write(f"{line}\n")
+
+
+def main():
+    parser = argparse.ArgumentParser(
+        description='Train an n-gram language model with KenLM to be used with BeamSearch decoder of ASR models.'
+    )
+    parser.add_argument("--input_path", required=True, type=str)
+    parser.add_argument("--model_path", required=True, type=str)
+    parser.add_argument("--lm_model", required=True, type=str)
+    parser.add_argument("--ngram_length", required=True, type=int)
+    parser.add_argument("--kenlm_path", required=True, type=str)
+    parser.add_argument("--do_lowercase", action='store_true')
+    #parser.add_argument("--processed_text_file", required=True, type=str)
+    args = parser.parse_args()
+
+    """ LMPLZ ARGUMENT SETUP """
+    args = [
+        os.path.join(args.kenlm_path, 'lmplz'),
+        "-o",
+        args.ngram_length,
+        "--text",
+        args.input_path,
+        "--arpa",
+        f"{args.lm_model}.tmp.arpa",
+        "--discount_fallback",
+    ]
+
+    result = subprocess.run(args, capture_output=False, text=True, stdout=sys.stdout, stderr=sys.stderr)
+
+    """ BINARY BUILD """
+    args = [os.path.join(args.kenlm_path, "build_binary"), "trie", f"{args.lm_model}.tmp.arpa", args.lm_model]
+
+    logging.info(f"Running binary_build command \n\n{' '.join(args)}\n\n")
+
+    result = subprocess.run(args, capture_output=False, text=True, stdout=sys.stdout, stderr=sys.stderr)
+
+
+if __name__ == '__main__':
+    main()
